Update run_docker.sh to start container attached
Per project requirements, container needs to be attached
to STDIN to view the log output from the application.
The script will now build a fresh image, run in detached
mode to output information about the container, then
attach to STDIN.
It can then be exited with CTRL+C which will stop the container
and then remove the container so it can be rebuilt and run
using the run_docker script again.